refactor: manager actions

+++ src/router/v1/service/deleteRoute.ts
@@ -1,6 +1,7 @@
-import {AppContext} from "../../../app";
+import {AppContext} from "@nsm/app";
 import {RouterHandler} from "../../index";
 import {handleErr} from "@nsm/util/routes";
+import {checkServiceExists} from "@nsm/router/util/preconditions";
 
 export default async function ({manager}: AppContext): Promise<RouterHandler> {
     return {
@@ -12,13 +13,13 @@ export default async function ({manager}: AppContext): Promise<RouterHandler> {
                     res.status(400).json({status: 400, message: 'Required \'id\' field not present in the body.'});
                     return;
                 }
+                if (!await checkServiceExists(id, manager, res)) {
+                    return;
+                }
                 try {
-                    const result = await manager.deleteService(id);
-                    if (result) {
-                        res.status(200).json({status: 200, message: 'Service deleted.'});
-                    } else {
-                        res.status(404).json({status: 404, message: 'Service not found or unknown error occured.'});
-                    }
+                    await manager.deleteService(id);
+
+                    res.status(200).json({status: 200, message: 'Service deleted.'});
                 } catch (e) {
                     handleErr(e, res);
                 }
